xxx-FabricRealTimeAnalytics/Coach/Resources/Docker Files/requirements.txt
azure-core<=1.28.0
azure-eventhub<=5.11.3
certifi<=2023.5.7
charset-normalizer<=3.2.0
idna<=3.4
numpy<=1.25.1
pandas<=2.0.3
python-dateutil<=2.8.2
pytz<=2023.3
requests<=2.31.0
six<=1.16.0
tabulate<=0.9.0
typing_extensions<=4.7.1
tzdata<=2023.3
urllib3<=2.0.4